Merges 128 commits from dev including: FCMP++ curve-tree integration,
hybrid PQC KEM scanning, shekyl-cli full parity, shekyl-address
Bech32m encoding, native Rust transaction signing, staking
enhancements, wallet/api removal, and ZeroMQ cleanup.

Conflict resolution: docs/MERKLE_TREE.md — took dev version which has
the polished, structured version of the document. The main-only version
was an earlier draft.

Made-with: Cursor

+++ src/common/utf8.h
@@ -85,7 +85,7 @@ namespace tools
       else
         throw std::runtime_error("Invalid UTF-8");
 
-      cp = t(cp);
+      cp = static_cast<uint32_t>(t(cp));
       if (cp <= 0x7f)
         bytes = 1;
       else if (cp <= 0x7ff)
